refactor(kolme): extract adapter-backed client module ownership (#1950)
