+++ crates/kamn-core/src/kolme_runtime_commit.rs
@@ -110,6 +110,7 @@ use std::process::{Command, Stdio};
 use std::thread;
 use std::time::{Duration, Instant};
 
+mod adapter_backed_client;
 mod block_fallback_reconciler;
 mod finality_checker;
 mod fork_finality_resolver;
@@ -505,6 +506,7 @@ pub enum KolmeRuntimeCommitOutcome {
 /// Runtime lifecycle state projected from commit receipt outcomes.
 pub type RuntimeCommitLifecycleState = KamnKolmeRuntimeCommitLifecycleState;
 
+pub use adapter_backed_client::AdapterBackedKolmeRuntimeCommitClient;
 pub use block_fallback_reconciler::KolmeRuntimeCommitBlockFallbackReconciler;
 pub use finality_checker::KolmeRuntimeCommitFinalityChecker;
 pub use fork_finality_resolver::KolmeRuntimeCommitForkFinalityResolver;
@@ -1201,108 +1203,6 @@ impl<T: KolmeRuntimeCommitProviderTransport> KolmeRuntimeCommitProvider
                     reason: error.to_string(),
                 })?;
         Ok(outcome.into())
-    }
-}
-
-/// Adapter-backed runtime commit client that enforces provider and finality policies.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct AdapterBackedKolmeRuntimeCommitClient<P> {
-    expected_provider: String,
-    provider: P,
-}
-
-impl<P: KolmeRuntimeCommitProvider> AdapterBackedKolmeRuntimeCommitClient<P> {
-    /// Builds adapter-backed client with expected provider identifier.
-    pub fn new(expected_provider: &str, provider: P) -> Result<Self, KolmeRuntimeCommitError> {
-        if !is_kolme_valid_expected_provider_input_contract(expected_provider) {
-            return Err(KolmeRuntimeCommitError::InvalidRequest {
-                field: "expected_provider",
-                reason: "must not be empty",
-            });
-        }
-        Ok(Self {
-            expected_provider: expected_provider.to_owned(),
-            provider,
-        })
-    }
-}
-
-impl<P: KolmeRuntimeCommitProvider> KolmeRuntimeCommitClient
-    for AdapterBackedKolmeRuntimeCommitClient<P>
-{
-    fn submit_commit(
-        &mut self,
-        request: &KolmeRuntimeCommitRequest,
-    ) -> Result<KolmeRuntimeCommitOutcome, KolmeRuntimeCommitError> {
-        request.validate()?;
-        let expected_provider = self.expected_provider.as_str();
-        let map_provider_receipt = |receipt: KolmeRuntimeCommitProviderReceipt| {
-            validate_kolme_provider_receipt_identity_contract(
-                expected_provider,
-                receipt.provider.as_str(),
-                receipt.commit_id.as_str(),
-            )
-            .map_err(|error| match error {
-                KamnKolmeProviderReceiptIdentityError::ProviderMismatch { expected, observed } => {
-                    KolmeRuntimeCommitError::ProviderMismatch { expected, observed }
-                }
-                KamnKolmeProviderReceiptIdentityError::EmptyCommitId => {
-                    KolmeRuntimeCommitError::InvalidRequest {
-                        field: "receipt_commit_id",
-                        reason: "must not be empty",
-                    }
-                }
-            })?;
-            require_kolme_final_receipt_finality_contract(receipt.finality).map_err(|error| {
-                match error {
-                    KamnKolmeRuntimeLifecyclePolicyError::NonFinalReceipt { finality } => {
-                        KolmeRuntimeCommitError::NonFinalReceipt {
-                            commit_id: receipt.commit_id.clone(),
-                            finality,
-                        }
-                    }
-                }
-            })?;
-            Ok(KolmeRuntimeCommitReceipt {
-                provider: receipt.provider,
-                commit_id: receipt.commit_id,
-                finality: receipt.finality,
-            })
-        };
-        let provider_outcome = self
-            .provider
-            .submit_runtime_commit(&request.to_wire_payload(), request.idempotency_key())
-            .map_err(|error| match error {
-                KolmeRuntimeCommitProviderError::Timeout => {
-                    KolmeRuntimeCommitError::ProviderTransport {
-                        kind: KolmeRuntimeCommitTransportErrorKind::Timeout,
-                        detail: "provider request timed out".to_owned(),
-                    }
-                }
-                KolmeRuntimeCommitProviderError::Unavailable { reason } => {
-                    KolmeRuntimeCommitError::ProviderTransport {
-                        kind: KolmeRuntimeCommitTransportErrorKind::Unavailable,
-                        detail: reason,
-                    }
-                }
-                KolmeRuntimeCommitProviderError::MalformedResponse { reason } => {
-                    KolmeRuntimeCommitError::ProviderTransport {
-                        kind: KolmeRuntimeCommitTransportErrorKind::MalformedResponse,
-                        detail: reason,
-                    }
-                }
-            })?;
-        match provider_outcome {
-            KolmeRuntimeCommitProviderOutcome::Submitted(receipt) => Ok(
-                KolmeRuntimeCommitOutcome::Submitted(map_provider_receipt(receipt)?),
-            ),
-            KolmeRuntimeCommitProviderOutcome::Duplicate(receipt) => Ok(
-                KolmeRuntimeCommitOutcome::Duplicate(map_provider_receipt(receipt)?),
-            ),
-            KolmeRuntimeCommitProviderOutcome::Rejected { reason } => {
-                Ok(KolmeRuntimeCommitOutcome::Rejected { reason })
-            }
-        }
     }
 }
